Move legend to separate component
