+++ app/components/map-utility-box.js
@@ -1,5 +1,47 @@
 import Component from '@ember/component';
+import { computed } from 'ember-decorators/object';
+import numeral from 'numeral';
 
 export default Component.extend({
   handleGeographyLevelToggle() {},
+
+  @computed('mapConfig')
+  breaks(mapConfig) {
+    // return an array of objects, each with a display-ready range and color
+    const { layers = [] } = mapConfig;
+    console.log(layers);
+    const [firstLayer = {}] = layers;
+    const { paintConfig: config = {} } = firstLayer;
+    const { isPercent, breaks = [], colors = [] } = config;
+    console.log(breaks);
+    const format = (value) => { // eslint-disable-line
+      return isPercent ? numeral(value).format('0,0%') : numeral(value).format('0,0');
+    };
+
+    const breaksArray = [];
+
+    for (let i = breaks.length - 1; i >= 0; i -= 1) {
+      if (i === breaks.length - 1) {
+        breaksArray.push({
+          label: `${format(breaks[breaks.length - 2])} or more`,
+          color: colors[breaks.length - 1],
+        });
+        continue; // eslint-disable-line
+      }
+
+      if (i === 0) {
+        breaksArray.push({
+          label: isPercent ? `Less than ${format(breaks[0])}` : `Under ${format(breaks[0])}`,
+          color: colors[0],
+        });
+        continue; // eslint-disable-line
+      }
+
+      breaksArray.push({
+        label: `${format(breaks[i - 1])} - ${format(breaks[i])}`,
+        color: colors[i],
+      });
+    }
+    return breaksArray;
+  },
 });
