+++ app/templates/components/map-utility-box.hbs
@@ -8,6 +8,8 @@
   {{/each}}
 </div>
 
+{{log breaks}}
+
 <ul class="menu geography-selector">
   {{#each toggles as |toggle|}}
     <li class="{{if (eq geographyLevel toggle.type) 'active'}} explode-block toggle-list-item">
